comp3900/backend/APIbackend/requirements.txt
Pillow==8.2.0
markdown==3.3.4
django==3.2.4
djangorestframework==3.12.4
django-rest-passwordreset==1.2.0
django-friendship==1.9.1
djangochannelsrestframework==0.3.0
django-apscheduler==0.6.0
torch==1.9.0